Do not export class definitions. Export only types and function that creates Registry singleton.
