+++ src/index.ts
@@ -1,9 +1,5 @@
 import { Registry } from './registry';
 
-export * from './registry';
-export * from './collector';
-export * from './counter';
-export * from './gauge';
-export * from './histogram';
+export * from './types';
 
 export default () => new Registry();
